+++ sdk/nodejs/eventhub/zMixins.ts
@@ -13,14 +13,14 @@
 // limitations under the License.
 
 import * as pulumi from "@pulumi/pulumi";
-import { getServiceBusNamespace, Subscription, Topic } from ".";
+import { getServiceBusNamespace, Queue, Subscription, Topic } from ".";
 import { EventHub, EventHubConsumerGroup, getEventhubNamespace } from ".";
 
 import * as appservice from "../appservice";
 
-interface TopicBindingDefinition extends appservice.BindingDefinition {
+interface ServiceBusBindingDefinition extends appservice.BindingDefinition {
     /**
-     * The name of the property in the context object to bind the actual topic message to.
+     * The name of the property in the context object to bind the actual ServiceBus message to.
      */
     name: string;
 
@@ -30,19 +30,24 @@ interface TopicBindingDefinition extends appservice.BindingDefinition {
     type: "serviceBusTrigger";
 
     /**
-     * The direction of the binding.  We only support topics being inputs to functions.
+     * The direction of the binding.  We only support queues and topics being inputs to functions.
      */
     direction: "in";
 
     /**
-     * The name of the topic we are subscribing to.
+     * Name of the queue to monitor. Set only if monitoring a queue, not for a topic.
      */
-    topicName: pulumi.Input<string>;
+    queueName?: pulumi.Input<string>;
 
     /**
-     * The name of the subscription inside the topic we are subscribing to.
+     * Name of the topic to monitor. Set only if monitoring a topic, not for a queue.
      */
-    subscriptionName: pulumi.Input<string>;
+    topicName?: pulumi.Input<string>;
+
+    /**
+     * Name of the subscription to monitor. Set only if monitoring a topic, not for a queue.
+     */
+    subscriptionName?: pulumi.Input<string>;
 
     /**
      * The name of an app setting that contains the Service Bus connection string to use for this binding.
@@ -51,16 +56,16 @@ interface TopicBindingDefinition extends appservice.BindingDefinition {
 }
 
 /**
- * Data that will be passed along in the context object to the TopicCallback.
+ * Data that will be passed along in the context object to the ServiceBusCallback.
  */
-export interface TopicContext extends appservice.Context<void> {
+export interface ServiceBusContext extends appservice.Context<void> {
     invocationId: string;
     executionContext: {
         invocationId: string;
         functionName: string;
         functionDirectory: string;
     };
-    bindings: { topic: string };
+    bindings: { message: string };
     bindingData: {
         deadLetterSource: any;
         messageId: string;
@@ -70,7 +75,7 @@ export interface TopicContext extends appservice.Context<void> {
         label: any;
         correlationId: any,
         properties: Record<string, any>;
-        topic: string,
+        message: string,
         sys: {
             methodName: string;
             utcNow: string;
@@ -80,7 +85,7 @@ export interface TopicContext extends appservice.Context<void> {
 }
 
 /**
- * Host settings specific to the Service Bus Topic/Subscription plugin.
+ * Host settings specific to the Service Bus Queue/Topic/Subscription plugin.
  *
  * For more details see https://docs.microsoft.com/en-us/azure/azure-functions/functions-bindings-service-bus#host-json
  */
@@ -102,29 +107,40 @@ export interface ServiceBusHostExtensions extends appservice.HostSettings {
         maxAutoRenewDuration?: string,
     },
 }
-export interface TopicHostSettings extends appservice.HostSettings {
+export interface ServiceBusHostSettings extends appservice.HostSettings {
     extensions?: {
         serviceBus: ServiceBusHostExtensions,
     }
 }
 
 /**
- * Signature of the callback that can receive topic notifications.
+ * Signature of the callback that can receive queue and topic notifications.
  */
-export type TopicCallback = appservice.Callback<TopicContext, string, void>;
+export type ServiceBusCallback = appservice.Callback<ServiceBusContext, string, void>;
 
-export interface TopicEventSubscriptionArgs extends appservice.CallbackFunctionAppArgs<TopicContext, any, void> {
+export interface ServiceBusFunctionArgs extends appservice.CallbackArgs<ServiceBusContext, string, void> {
     /**
-     * The name of the resource group in which to create the event subscription. [resourceGroup] takes precedence over [resourceGroupName].
-     * If none of the two is supplied, the Topic's resource group will be used.
+     * The ServiceBus Queue to subscribe to.
      */
-    resourceGroupName?: pulumi.Input<string>;
+    queue?: Queue;
 
     /**
-     * The Subscription to subscribe the FunctionApp to.  If not present, a new Subscription
-     * resource will be created.
+     * The ServiceBus Topic to subscribe to.
+     */
+    topic?: Topic;
+
+    /**
+     * The ServiceBus Subscription to subscribe the Function to.
      */
     subscription?: Subscription;
+};
+
+export interface QueueEventSubscriptionArgs extends appservice.CallbackFunctionAppArgs<ServiceBusContext, string, void> {
+    /**
+     * The name of the resource group in which to create the event subscription. [resourceGroup] takes precedence over [resourceGroupName].
+     * If none of the two is supplied, the Queue's resource group will be used.
+     */
+    resourceGroupName?: pulumi.Input<string>;
 
     /**
      * The maximum number of deliveries.  Will default to 10 if not specified.
@@ -135,8 +151,54 @@ export interface TopicEventSubscriptionArgs extends appservice.CallbackFunctionA
      * Host settings specific to the Service Bus Topic/Subscription plugin. These values can be provided here, or defaults will
      * be used in their place.
      */
-    hostSettings?: TopicHostSettings;
+    hostSettings?: ServiceBusHostSettings;
 };
+
+export interface TopicEventSubscriptionArgs extends QueueEventSubscriptionArgs {
+    /**
+     * The Subscription to subscribe the FunctionApp to.  If not present, a new Subscription
+     * resource will be created.
+     */
+    subscription?: Subscription;
+};
+
+declare module "./queue" {
+    interface Queue {
+        /**
+         * Creates a new subscription to events fired from this Queue to the handler provided, along
+         * with options to control the behavior of the subscription.
+         */
+        onEvent(
+            name: string, args: ServiceBusCallback | QueueEventSubscriptionArgs, opts?: pulumi.ComponentResourceOptions): QueueEventSubscription;
+    }
+}
+
+Queue.prototype.onEvent = function(this: Queue, name, args, opts) {
+    const functionArgs = args instanceof Function
+        ? <QueueEventSubscriptionArgs>{ callback: args }
+        : args;
+
+    return new QueueEventSubscription(name, this, functionArgs, opts);
+}
+
+export class QueueEventSubscription extends appservice.EventSubscription<ServiceBusContext, string, void> {
+    readonly queue: Queue;
+
+    constructor(
+        name: string, queue: Queue,
+        args: QueueEventSubscriptionArgs, opts: pulumi.ComponentResourceOptions = {}) {
+        const { resourceGroupName, location } = appservice.getResourceGroupNameAndLocation(args, queue.resourceGroupName);
+
+        super("azure:eventhub:QueueEventSubscription",
+            name,
+            new ServiceBusFunction(name, { queue }),
+            { ...args, resourceGroupName, location },
+            { parent: queue, ...opts });
+
+        this.queue = queue;
+        this.registerOutputs();
+    }
+}
 
 declare module "./topic" {
     interface Topic {
@@ -145,7 +207,7 @@ declare module "./topic" {
          * with options to control the behavior of the subscription.
          */
         onEvent(
-            name: string, args: TopicCallback | TopicEventSubscriptionArgs, opts?: pulumi.ComponentResourceOptions): TopicEventSubscription;
+            name: string, args: ServiceBusCallback | TopicEventSubscriptionArgs, opts?: pulumi.ComponentResourceOptions): TopicEventSubscription;
     }
 }
 
@@ -157,7 +219,7 @@ Topic.prototype.onEvent = function(this: Topic, name, args, opts) {
     return new TopicEventSubscription(name, this, functionArgs, opts);
 }
 
-export class TopicEventSubscription extends appservice.EventSubscription<TopicContext, string, void> {
+export class TopicEventSubscription extends appservice.EventSubscription<ServiceBusContext, string, void> {
     readonly topic: Topic;
     readonly subscription: Subscription;
 
@@ -176,41 +238,61 @@ export class TopicEventSubscription extends appservice.EventSubscription<TopicCo
             maxDeliveryCount: pulumi.output(args.maxDeliveryCount).apply(c => c === undefined ? 10 : c),
         }, opts);
 
-        // The topic binding does not store the Service Bus connection string directly.  Instead, the
-        // connection string is put into the app settings (under whatever key we want). Then, the
-        // .connection property of the binding contains the *name* of that app setting key.
-        const bindingConnectionKey = "BindingConnectionAppSettingsKey";
-
-        const bindings: TopicBindingDefinition[] = [{
-            name: "topic",
-            direction: "in",
-            type: "serviceBusTrigger",
-            topicName: topic.name,
-            subscriptionName: subscription.name,
-            connection: bindingConnectionKey,
-        }];
-
-        const namespace = pulumi.all([topic.namespaceName, topic.resourceGroupName])
-                               .apply(([namespaceName, resourceGroupName]) =>
-                                    getServiceBusNamespace({ name: namespaceName, resourceGroupName }));
-
-        // Place the mapping from the well known key name to the Service Bus account connection string in
-        // the 'app settings' object.
-
-        const appSettings = pulumi.all([args.appSettings, namespace.defaultPrimaryConnectionString]).apply(
-            ([appSettings, connectionString]) => ({ ...appSettings, [bindingConnectionKey]: connectionString }));
-
-        super("azure:eventhub:TopicEventSubscription", name, bindings, {
-            ...args,
-            resourceGroupName,
-            location,
-            appSettings
-        }, opts);
+        super("azure:eventhub:TopicEventSubscription",
+            name,
+            new ServiceBusFunction(name, { topic, subscription }),
+            { ...args, resourceGroupName, location },
+            opts);
 
         this.topic = topic;
         this.subscription = subscription;
 
         this.registerOutputs();
+    }
+}
+
+/**
+ * Azure Function triggered by a ServiceBus Topic.
+ */
+export class ServiceBusFunction extends appservice.Function<ServiceBusContext, string, void> {
+    constructor(name: string, args: ServiceBusFunctionArgs) {
+        if (!args.queue && !args.topic) {
+            throw new Error("Either [queue] or [topic] has to be specified");
+        }
+
+        if (args.queue && args.topic) {
+            throw new Error("Either one of [queue] or [topic] has to be specified, not both");
+        }
+
+        if (args.queue && args.subscription) {
+            throw new Error("[subscription] can't be specified in combination with [queue]");
+        }
+
+        const namespaceName = (args.queue && args.queue.namespaceName) || args.topic!.namespaceName;
+        const resourceGroupName = (args.queue && args.queue.resourceGroupName) || args.topic!.resourceGroupName;
+
+        // The topic binding does not store the Service Bus connection string directly.  Instead, the
+        // connection string is put into the app settings (under whatever key we want). Then, the
+        // .connection property of the binding contains the *name* of that app setting key.
+        const bindingConnectionKey = "BindingConnectionAppSettingsKey";
+
+        const bindings: ServiceBusBindingDefinition[] = [{
+            name: "message",
+            direction: "in",
+            type: "serviceBusTrigger",
+            queueName: args.queue && args.queue.name,
+            topicName: args.topic && args.topic.name,
+            subscriptionName: args.subscription && args.subscription.name,
+            connection: bindingConnectionKey,
+        }];
+
+        const namespace = pulumi.all([namespaceName, resourceGroupName])
+                               .apply(([namespaceName, resourceGroupName]) =>
+                                    getServiceBusNamespace({ name: namespaceName, resourceGroupName }));
+
+        const appSettings = namespace.defaultPrimaryConnectionString.apply(connectionString => ({ [bindingConnectionKey]: connectionString }));
+
+        super(name, bindings, args, appSettings);
     }
 }
 
@@ -297,6 +379,23 @@ export interface EventHubContext extends appservice.Context<void> {
  */
 export type EventHubCallback = appservice.Callback<EventHubContext, string, void>;
 
+export interface EventHubFunctionArgs extends appservice.CallbackArgs<EventHubContext, any, void> {
+    /**
+     * Event Hub to subscribe the Function to.
+     */
+    eventHub: EventHub;
+
+    /**
+     * Optional Consumer Group to subscribe the FunctionApp to. If not present, the default consumer group will be used.
+     */
+    consumerGroup?: EventHubConsumerGroup;
+
+    /**
+     * Set to 'many' in order to enable batching. If omitted or set to 'one', single message passed to function.
+     */
+    cardinality?: pulumi.Input<"many" | "one">;
+};
+
 export interface EventHubSubscriptionArgs extends appservice.CallbackFunctionAppArgs<EventHubContext, any, void> {
     /**
      * The name of the resource group in which to create the event subscription. [resourceGroup] takes precedence over [resourceGroupName].
@@ -336,7 +435,7 @@ EventHub.prototype.onEvent = function(this: EventHub, name, args, opts) {
 
 export class EventHubSubscription extends appservice.EventSubscription<EventHubContext, string, void> {
     readonly eventHub: EventHub;
-    readonly consumerGroup: EventHubConsumerGroup;
+    readonly consumerGroup?: EventHubConsumerGroup;
 
     constructor(
         name: string, eventHub: EventHub,
@@ -346,12 +445,26 @@ export class EventHubSubscription extends appservice.EventSubscription<EventHubC
 
         const { resourceGroupName, location } = appservice.getResourceGroupNameAndLocation(args, eventHub.resourceGroupName);
 
-        const consumerGroup = args.consumerGroup || new EventHubConsumerGroup(name, {
-            resourceGroupName,
-            namespaceName: eventHub.namespaceName,
-            eventhubName: eventHub.name,
-        }, opts);
+        super("azure:eventhub:EventHubSubscription",
+            name,
+            new EventHubFunction(name, { ...args, eventHub }),
+            { ...args, resourceGroupName, location },
+            opts);
 
+        this.eventHub = eventHub;
+        this.consumerGroup = args.consumerGroup;
+
+        this.registerOutputs();
+    }
+}
+
+export const DefaultConsumerGroup = "$Default";
+
+/**
+ * Azure Function triggered by an Event Hub.
+ */
+export class EventHubFunction extends appservice.Function<EventHubContext, string, void> {
+    constructor(name: string, args: EventHubFunctionArgs) {
         // The event hub binding does not store the Event Hubs connection string directly.  Instead, the
         // connection string is put into the app settings (under whatever key we want). Then, the
         // .connection property of the binding contains the *name* of that app setting key.
@@ -361,32 +474,21 @@ export class EventHubSubscription extends appservice.EventSubscription<EventHubC
             name: "eventHub",
             direction: "in",
             type: "eventHubTrigger",
-            eventHubName: eventHub.name,
-            consumerGroup: consumerGroup.name,
+            eventHubName: args.eventHub.name,
+            consumerGroup: args.consumerGroup !== undefined ? args.consumerGroup.name : DefaultConsumerGroup,
             cardinality: args.cardinality,
             connection: bindingConnectionKey,
         }];
 
-        const namespace = pulumi.all([eventHub.namespaceName, eventHub.resourceGroupName])
+        const namespace = pulumi.all([args.eventHub.namespaceName, args.eventHub.resourceGroupName])
                                .apply(([namespaceName, resourceGroupName]) =>
                                     getEventhubNamespace({ name: namespaceName, resourceGroupName }));
 
         // Place the mapping from the well known key name to the Event Hubs account connection string in
         // the 'app settings' object.
 
-        const appSettings = pulumi.all([args.appSettings, namespace.defaultPrimaryConnectionString]).apply(
-            ([appSettings, connectionString]) => ({ ...appSettings, [bindingConnectionKey]: connectionString }));
+        const appSettings = namespace.defaultPrimaryConnectionString.apply(connectionString => ({ [bindingConnectionKey]: connectionString }));
 
-        super("azure:eventhub:EventHubSubscription", name, bindings, {
-            ...args,
-            resourceGroupName,
-            location,
-            appSettings
-        }, opts);
-
-        this.eventHub = eventHub;
-        this.consumerGroup = consumerGroup;
-
-        this.registerOutputs();
+        super(name, bindings, args, appSettings);
     }
 }